chore: run rekres (update to Go 1.22)
Update `handlers` function to match new `http.ServeMux` internals.
