+++ .dockerignore
@@ -1,12 +1,14 @@
 # THIS FILE WAS AUTOMATICALLY GENERATED, PLEASE DO NOT EDIT.
 #
-# Generated on 2023-07-10T16:20:59Z by kres latest.
+# Generated on 2024-02-08T22:38:02Z by kres latest.
 
 *
 !debug.go
 !debug_off.go
 !debug_off_test.go
 !debug_on.go
+!debug_on_121.go
+!debug_on_122.go
 !debug_on_test.go
 !race_off.go
 !race_on.go
