+++ go.mod
@@ -1,10 +1,10 @@
 module github.com/siderolabs/go-debug
 
-go 1.20
+go 1.22.0
 
 require (
 	github.com/stretchr/testify v1.8.4
-	golang.org/x/net v0.12.0
+	golang.org/x/net v0.21.0
 )
 
 require (
